third_party/WebKit/LayoutTests/fast/dom/shadow/getComputedStyle-flat-tree-parent-dirty-expected.txt

CONSOLE WARNING: line 12: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
getComputedStyle should update style if the parent node in the flat tree needs a recalc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inside).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
